requirements.txt
google-generativeai
google-ads
